apps/github/src/inngest/functions/users/schedule-users-syncs.ts
import { env } from '@/env';
import { db } from '@/database/client';
import { Organisation } from '@/database/schema';
import { inngest } from '../../client';

export const scheduleUsersSyncs = inngest.createFunction(
  { id: 'schedule-users-syncs' },
  { cron: env.USERS_SYNC_CRON },
  async ({ step }) => {
    const organisations = await db
      .select({
        id: Organisation.id,
        installationId: Organisation.installationId,
        accountLogin: Organisation.accountLogin,
        region: Organisation.region,
      })
      .from(Organisation);

    if (organisations.length > 0) {
      await step.sendEvent(
        'sync-organisations-users',
        organisations.map(({ id, installationId, accountLogin, region }) => ({
          name: 'users/page_sync.requested',
          data: {
            installationId,
            organisationId: id,
            region,
            accountLogin,
            cursor: null,
            syncStartedAt: Date.now(),
            isFirstSync: false,
          },
        }))
      );
    }

    return { organisations };
  }
);
